+++ src/conf/default_conf.rs
@@ -214,6 +214,14 @@ leave_broot = false
 # help_code = "gray(21) gray(3)"
 # help_headers = "ansi(208) None"
 # help_table_border = "ansi(239) None"
+# preview = "gray(20) gray(1) / gray(18) gray(2)"
+# preview_line_number = "gray(12) gray(3)"
+# preview_match = "None ansi(29)"
+# hex_null = "gray(8) None"
+# hex_ascii_graphic = "gray(18) None"
+# hex_ascii_whitespace = "ansi(143) None"
+# hex_ascii_other = "ansi(215) None"
+# hex_non_ascii = "ansi(167) None"
 
 # You may find explanations and other skins on
 #  https://dystroy.org/broot/skins