+++ src/errors.rs
@@ -32,7 +32,7 @@ custom_error! {pub TreeBuildError
 
 custom_error! {pub ConfError
     Io {source: io::Error}                          = "unable to read from the file: {}",
-    Toml {source: toml::de::Error}                  = "unable to parse TOML",
+    Toml {source: toml::de::Error}                  = "unable to parse TOML: {}",
     MissingField {txt: String}                      = "missing field in conf",
     InvalidVerbInvocation {invocation: String}      = "invalid verb invocation: {}",
     UnknownInternal {verb: String}                  = "not a known internal: {}",