+++ src/pattern/regex_patterns.rs
@@ -15,7 +15,7 @@ pub struct RegexPattern {
 
 impl fmt::Display for RegexPattern {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}", self.rex, self.flags)
+        write!(f, "/{}/{}", self.rex, self.flags)
     }
 }