+++ src/skin/style_map.rs
@@ -167,7 +167,7 @@ StyleMap! {
     help_table_border: ansi(239), None, []
     preview: gray(20), gray(1), [] / gray(18), gray(2), []
     preview_line_number: gray(12), gray(3), []
-    preview_match: None, ansi(29), [Underlined]
+    preview_match: None, ansi(29), []
     hex_null: gray(8), None, []
     hex_ascii_graphic: gray(18), None, []
     hex_ascii_whitespace: ansi(143), None, []