+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - selected line in preview, interesting when removing the pattern (to see what's around a match)
 - faster availability of huge files in preview
 - search in preview now interrupted by key events (just like the trees)
+- a content search in a tree is propagated as a regex in a preview on :panel_right (ctrl-right)
 
 <a name="v0.19.1"></a>
 ### v0.19.1 - 2020-07-17